signs are fixed in continous gauge fixing procedure for arrays (i.e. Z2 broken tensors).
In differentiability test added an option to turn off
discete gauge fixing. This does not change the result
as all signs in eigenvalue decomposition and svd are fixed.

In eigensystem.jl added an option to compute Z2 odd
sector eigenvalues. Feature does not work yet. To be
fixed soon.

In GiltTNR2D_essentials.py - minor correction.